+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 Used [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.19.0] - 2021-10-12
+### Added
+- Methods beneficiaryCommissionList, beneficiaryCommissionAdd, beneficiaryCommissionEdit, beneficiaryCommissionDelete
+### Removed
+- Parameter $commissionProc in methods beneficiaryAdd
+
 ## [v1.18.0] - 2021-10-08
 ### Added
 - Methods beneficiaryList, beneficiaryAdd, beneficiaryEdit, beneficiaryBlock, beneficiaryUnblock, beneficiaryShow